home project/correspondence_management/requirements.txt
# متطلبات برنامج إدارة المراسلات
# Correspondence Management System Requirements

# المتطلبات الأساسية
matplotlib>=3.5.0
numpy>=1.21.0
Pillow>=8.3.0

# متطلبات إضافية (اختيارية)
# openpyxl>=3.0.9          # لتصدير Excel
# python-dateutil>=2.8.2   # لمعالجة التواريخ المتقدمة
# tkcalendar>=1.6.1        # لاختيار التاريخ بصرياً
# arabic-reshaper>=2.1.3   # لتحسين النصوص العربية
# python-bidi>=0.4.2       # لدعم النصوص ثنائية الاتجاه

# ملاحظة: tkinter و sqlite3 مدمجان مع Python